# docs/openapi.yaml
components:
  schemas:
...
    AuthResponse:
      properties:
        auth_date:
          title: Auth Date
          type: integer
        user:
          $ref: '#/components/schemas/UserView'
      required:
      - user
      - auth_date
      title: AuthResponse
      type: object
    BootstrapResponse:
      properties:
        relationships:
          items:
            $ref: '#/components/schemas/RelationshipView'
          title: Relationships
          type: array
        user:
          $ref: '#/components/schemas/UserView'
      required:
      - user
      - relationships
      title: BootstrapResponse
      type: object
    ConsentRequiredError:
      additionalProperties: false
      description: '403 body: the Telegram user has no current personal-data consent. Consent is given in the bot via /start.'
      properties:
        bot_command:
          const: /start
          default: /start
          title: Bot Command
          type: string
        error:
          const: consent_required
          default: consent_required
          title: Error
          type: string
        message:
          title: Message
          type: string
      required:
      - message
      title: ConsentRequiredError
      type: object
    DeliveryResponse:
      additionalProperties: false
      properties:
        request_id:
          format: uuid
          title: Request Id
          type: string
        status:
          enum:
          - ok
          - blocked
          - error
          title: Status
          type: string
        structured_result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Structured Result
        text:
          title: Text
          type: string
        version:
          default: '1.0'
          title: Version
          type: string
        workflow:
          $ref: '#/components/schemas/WorkflowName'
      required:
      - request_id
      - workflow
      - status
      - text
      title: DeliveryResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    MiniAppAssistBody:
      additionalProperties: false
      properties:
        language:
          anyOf:
          - maxLength: 16
            type: string
          - type: 'null'
          title: Language
        relationship_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Relationship Id
        text:
          maxLength: 10000
          minLength: 1
          title: Text
          type: string
      required:
      - text
      title: MiniAppAssistBody
      type: object
    RelationshipCreateBody:
      additionalProperties: false
      properties:
        aliases:
          items:
            type: string
          maxItems: 20
          title: Aliases
          type: array
        communication_style:
          additionalProperties: true
          title: Communication Style
          type: object
        relation_type:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Relation Type
        set_as_default:
          default: false
          title: Set As Default
          type: boolean
      title: RelationshipCreateBody
      type: object
    RelationshipRule:
      additionalProperties: false
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        priority:
          default: 0
          title: Priority
          type: integer
        type:
          title: Type
          type: string
        value:
          title: Value
          type: string
      required:
      - type
      - value
      title: RelationshipRule
      type: object
    RelationshipUpdateBody:
      additionalProperties: false
      properties:
        aliases:
          anyOf:
          - items:
              type: string
            maxItems: 20
            type: array
          - type: 'null'
          title: Aliases
        communication_style:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Communication Style
        relation_type:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Relation Type
      title: RelationshipUpdateBody
      type: object
    RelationshipView:
      additionalProperties: false
      properties:
        aliases:
          items:
            type: string
          title: Aliases
          type: array
        communication_style:
          additionalProperties: true
          title: Communication Style
          type: object
        relation_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Relation Type
        relationship_id:
          title: Relationship Id
          type: string
        rules:
          items:
            $ref: '#/components/schemas/RelationshipRule'
          title: Rules
          type: array
        ruleset_version:
          title: Ruleset Version
          type: integer
      required:
      - relationship_id
      - ruleset_version
      title: RelationshipView
      type: object
    RuleBody:
      additionalProperties: false
      properties:
        priority:
          default: 0
          maximum: 1000.0
          minimum: -1000.0
          title: Priority
          type: integer
        type:
          maxLength: 64
          minLength: 1
          title: Type
          type: string
        value:
          maxLength: 2000
          minLength: 1
          title: Value
          type: string
      required:
      - type
      - value
      title: RuleBody
      type: object
    UserView:
      additionalProperties: false
      properties:
        default_relationship_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Default Relationship Id
        first_name:
          title: First Name
          type: string
        language_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Language Code
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        telegram_user_id:
          title: Telegram User Id
          type: integer
        user_id:
          title: User Id
          type: string
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      required:
      - user_id
      - telegram_user_id
      - first_name
      title: UserView
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    WorkflowName:
      enum:
      - soften
      - decode
      - help-say
      title: WorkflowName
      type: string
info:
  title: Свои Правила API
  version: 0.6.0
openapi: 3.1.0
paths:
...
  /v1/miniapp/assist/{workflow}:
    post:
      operationId: miniapp_assist_v1_miniapp_assist__workflow__post
      parameters:
      - in: path
        name: workflow
        required: true
        schema:
          $ref: '#/components/schemas/WorkflowName'
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MiniAppAssistBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeliveryResponse'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Miniapp Assist
      tags:
      - miniapp
  /v1/miniapp/auth:
    post:
      operationId: miniapp_auth_v1_miniapp_auth_post
      parameters:
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Miniapp Auth
      tags:
      - miniapp
  /v1/miniapp/bootstrap:
    get:
      operationId: bootstrap_v1_miniapp_bootstrap_get
      parameters:
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Bootstrap
      tags:
      - miniapp
  /v1/miniapp/me/default-relationship/{relationship_id}:
    put:
      operationId: set_default_relationship_v1_miniapp_me_default_relationship__relationship_id__put
      parameters:
      - in: path
        name: relationship_id
        required: true
        schema:
          title: Relationship Id
          type: string
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserView'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Set Default Relationship
      tags:
      - miniapp
  /v1/miniapp/relationships:
    get:
      operationId: list_relationships_v1_miniapp_relationships_get
      parameters:
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RelationshipView'
                title: Response List Relationships V1 Miniapp Relationships Get
                type: array
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Relationships
      tags:
      - miniapp
    post:
      operationId: create_relationship_v1_miniapp_relationships_post
      parameters:
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipCreateBody'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipView'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Relationship
      tags:
      - miniapp
  /v1/miniapp/relationships/{relationship_id}:
    delete:
      operationId: delete_relationship_v1_miniapp_relationships__relationship_id__delete
      parameters:
      - in: path
        name: relationship_id
        required: true
        schema:
          title: Relationship Id
          type: string
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '204':
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Relationship
      tags:
      - miniapp
    patch:
      operationId: update_relationship_v1_miniapp_relationships__relationship_id__patch
      parameters:
      - in: path
        name: relationship_id
        required: true
        schema:
          title: Relationship Id
          type: string
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelationshipUpdateBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipView'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Relationship
      tags:
      - miniapp
  /v1/miniapp/relationships/{relationship_id}/rules:
    post:
      operationId: add_rule_v1_miniapp_relationships__relationship_id__rules_post
      parameters:
      - in: path
        name: relationship_id
        required: true
        schema:
          title: Relationship Id
          type: string
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleBody'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipRule'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Rule
      tags:
      - miniapp
  /v1/miniapp/relationships/{relationship_id}/rules/{rule_id}:
    delete:
      operationId: delete_rule_v1_miniapp_relationships__relationship_id__rules__rule_id__delete
      parameters:
      - in: path
        name: relationship_id
        required: true
        schema:
          title: Relationship Id
          type: string
      - in: path
        name: rule_id
        required: true
        schema:
          title: Rule Id
          type: integer
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '204':
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Rule
      tags:
      - miniapp
    put:
      operationId: update_rule_v1_miniapp_relationships__relationship_id__rules__rule_id__put
      parameters:
      - in: path
        name: relationship_id
        required: true
        schema:
          title: Relationship Id
          type: string
      - in: path
        name: rule_id
        required: true
        schema:
          title: Rule Id
          type: integer
      - in: header
        name: X-Telegram-Init-Data
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationshipRule'
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentRequiredError'
          description: Consent required
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Rule
      tags:
      - miniapp
...